Payments dataType changed to double

+++ payment-service/src/main/java/com/telran/paymentservice/paymentInformation/dto/PaymentInfoReturnDto.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -11,5 +12,5 @@ import java.util.UUID;
 @NoArgsConstructor
 public class PaymentInfoReturnDto {
     private UUID studentId;
-    private int paymentAmount;
+    private BigDecimal paymentAmount;
 }
